backend/requirements.txt
flask~=3.0.3
flask-cors~=5.0.1
flask-jwt-extended~=4.7.1
Flask-SQLAlchemy~=3.1.1
pymysql~=1.0.2
python-dotenv~=0.21.0
bcrypt~=3.2.0
numpy~=1.26.4
joblib~=1.4.2
pandas~=2.2.3
scapy~=2.6.1
scipy~=1.13.1
requests~=2.32.3
cicflowmeter~=0.2.0
pytest~=7.4.4
optuna~=4.3.0
scikit-learn~=1.6.1
imblearn~=0.0
imbalanced-learn~=0.12.3
torch==2.3.1+cpu
